+++ packable/packable/src/error.rs
@@ -3,7 +3,7 @@
 
 //! Errors related to packable operations.
 
-use core::convert::Infallible;
+use core::{convert::Infallible, fmt};
 
 mod sealed {
     use crate::error::UnpackError;
@@ -77,7 +77,7 @@ impl<T, U> From<U> for UnpackError<T, U> {
 impl<U> UnpackError<Infallible, U> {
     /// Get the [`Packer`](UnpackError::Unpacker) variant if the [`Packable`](UnpackError::Packable) variant is
     /// [`Infallible`].
-    pub fn into_unpacker(self) -> U {
+    pub fn into_unpacker_err(self) -> U {
         match self {
             Self::Packable(err) => match err {},
             Self::Unpacker(err) => err,
@@ -85,9 +85,53 @@ impl<U> UnpackError<Infallible, U> {
     }
 }
 
+impl<T> UnpackError<T, Infallible> {
+    /// Get the [`Packable`](UnpackError::Packable) variant if the [`Unpacker`](UnpackError::Unpacker) variant is
+    /// [`Infallible`].
+    pub fn into_packable_err(self) -> T {
+        match self {
+            Self::Packable(err) => err,
+            Self::Unpacker(err) => match err {},
+        }
+    }
+}
+
+impl<T, U> fmt::Display for UnpackError<T, U>
+where
+    T: fmt::Display,
+    U: fmt::Display,
+{
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Packable(err) => write!(f, "packable error while unpacking: {}", err),
+            Self::Unpacker(err) => write!(f, "unpacker error while unpacking: {}", err),
+        }
+    }
+}
+
+#[cfg(feature = "std")]
+impl<T, U> std::error::Error for UnpackError<T, U>
+where
+    T: std::error::Error,
+    U: std::error::Error,
+{
+}
+
+/// We cannot provide a [`From`] implementation because [`Infallible`] is not from this crate.
+#[allow(clippy::from_over_into)]
+impl Into<Infallible> for UnpackError<Infallible, Infallible> {
+    fn into(self) -> Infallible {
+        let (Self::Packable(err) | Self::Unpacker(err)) = self;
+        match err {}
+    }
+}
+
 /// Error type raised when an unknown tag is found while unpacking.
 #[derive(Debug)]
 pub struct UnknownTagError<T>(pub T);
+
+#[cfg(feature = "std")]
+impl<T> std::error::Error for UnknownTagError<T> where T: fmt::Display + fmt::Debug {}
 
 impl<T> From<Infallible> for UnknownTagError<T> {
     fn from(err: Infallible) -> Self {
@@ -95,7 +139,13 @@ impl<T> From<Infallible> for UnknownTagError<T> {
     }
 }
 
-/// Error type to be raised when [`&[u8]`] does not have enough bytes to unpack something or when
+impl<T: fmt::Display> fmt::Display for UnknownTagError<T> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "unknown tag value {}", self.0)
+    }
+}
+
+/// Error type to be raised when `&[u8]` does not have enough bytes to unpack something or when
 /// [`SlicePacker`]('crate::packer::SlicePacker') does not have enough space to pack something.
 #[derive(Debug)]
 pub struct UnexpectedEOF {
@@ -103,4 +153,17 @@ pub struct UnexpectedEOF {
     pub required: usize,
     /// The number of bytes the unpacker had or the number of bytes the packer can receive.
     pub had: usize,
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for UnexpectedEOF {}
+
+impl fmt::Display for UnexpectedEOF {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "not enough bytes, required {} but had {}",
+            self.required, self.had
+        )
+    }
 }