+++ packable/packable/src/packable/prefix.rs
@@ -16,37 +16,72 @@ use crate::{
 use alloc::{boxed::Box, vec::Vec};
 use core::{
     convert::Infallible,
-    fmt::Debug,
+    fmt,
     marker::PhantomData,
     ops::{Deref, DerefMut, Range},
 };
 
-/// Semantic error raised while unpacking a dynamically-sized sequences that use a type different than `usize` for their
-/// length-prefix.
+/// Semantic error raised while unpacking dynamically-sized sequences.
 #[derive(Debug)]
 pub enum UnpackPrefixError<T, E> {
-    /// Semantic error raised while unpacking an element of the sequence.
-    /// Typically this is [`Packable::UnpackError`](crate::Packable::UnpackError).
-    Packable(T),
+    /// Semantic error raised while unpacking an item of the sequence. Typically this is
+    /// [`Packable::UnpackError`](crate::Packable::UnpackError).
+    Item(T),
     /// Semantic error raised when the length prefix cannot be unpacked.
     Prefix(E),
 }
 
+impl<T: fmt::Display, E: fmt::Display> fmt::Display for UnpackPrefixError<T, E> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Item(err) => write!(f, "cannot unpack item: {}", err),
+            Self::Prefix(err) => write!(f, "cannot unpack prefix: {}", err),
+        }
+    }
+}
+
+#[cfg(feature = "std")]
+impl<T, E> std::error::Error for UnpackPrefixError<T, E>
+where
+    T: std::error::Error,
+    E: std::error::Error,
+{
+}
+
 impl<E> UnpackPrefixError<Infallible, E> {
     /// Projects the value to the [`Prefix`](UnpackPrefixError::Prefix) variant.
-    pub fn into_prefix(self) -> E {
+    pub fn into_prefix_err(self) -> E {
         match self {
-            Self::Packable(err) => match err {},
+            Self::Item(err) => match err {},
             Self::Prefix(err) => err,
         }
     }
 }
 
-impl<T, E> UnpackPrefixError<T, E> {
-    /// Returns the contained [`Packable`](UnpackPrefixError::Packable) value or computes it from a closure.
-    pub fn unwrap_packable_or_else<V: Into<T>>(self, f: impl FnOnce(E) -> V) -> T {
+impl<T> UnpackPrefixError<T, Infallible> {
+    /// Projects the value to the [`Item`](UnpackPrefixError::Item) variant.
+    pub fn into_item_err(self) -> T {
         match self {
-            Self::Packable(err) => err,
+            Self::Item(err) => err,
+            Self::Prefix(err) => match err {},
+        }
+    }
+}
+
+/// We cannot provide a [`From`] implementation because [`Infallible`] is not from this crate.
+#[allow(clippy::from_over_into)]
+impl Into<Infallible> for UnpackPrefixError<Infallible, Infallible> {
+    fn into(self) -> Infallible {
+        let (Self::Item(err) | Self::Prefix(err)) = self;
+        match err {}
+    }
+}
+
+impl<T, E> UnpackPrefixError<T, E> {
+    /// Returns the contained [`Item`](UnpackPrefixError::Item) value or computes it from a closure.
+    pub fn unwrap_item_err_or_else<V: Into<T>>(self, f: impl FnOnce(E) -> V) -> T {
+        match self {
+            Self::Item(err) => err,
             Self::Prefix(err) => f(err).into(),
         }
     }
@@ -54,7 +89,7 @@ impl<T, E> UnpackPrefixError<T, E> {
 
 impl<T, E> From<T> for UnpackPrefixError<T, E> {
     fn from(err: T) -> Self {
-        Self::Packable(err)
+        Self::Item(err)
     }
 }
 
@@ -114,8 +149,8 @@ impl<T, B> Packable for VecPrefix<T, B>
 where
     T: Packable,
     B: Bounded + Packable,
-    <B::Bounds as TryInto<B>>::Error: Debug,
-    <B as TryFrom<usize>>::Error: Debug,
+    <B::Bounds as TryInto<B>>::Error: fmt::Debug,
+    <B as TryFrom<usize>>::Error: fmt::Debug,
     Range<B::Bounds>: Iterator<Item = B::Bounds>,
 {
     type UnpackError = UnpackPrefixError<T::UnpackError, B::UnpackError>;
@@ -225,8 +260,8 @@ impl<T, B> Packable for BoxedSlicePrefix<T, B>
 where
     T: Packable,
     B: Bounded + Packable,
-    <B::Bounds as TryInto<B>>::Error: Debug,
-    <B as TryFrom<usize>>::Error: Debug,
+    <B::Bounds as TryInto<B>>::Error: fmt::Debug,
+    <B as TryFrom<usize>>::Error: fmt::Debug,
     Range<B::Bounds>: Iterator<Item = B::Bounds>,
 {
     type UnpackError = <VecPrefix<T, B> as Packable>::UnpackError;