+++ trace-tools/trace-tools/CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
-## 0.2.0 - 2021-01-31
+## 0.2.0 - 2021-02-18
 
 ### Added
 
@@ -27,7 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
- - `TraceSubscriber` type alias and `SubscriberBuilder` type now accomodate for the new layer;
+ - `TraceSubscriber` type alias and `SubscriberBuilder` type now accomodates for the new layer;
+ - Updated dependencies;
 
 ## 0.1.0 - 2021-01-21